fix test scripts

+++ scripts/test_bench_goblint_success.sh
@@ -5,7 +5,7 @@ set -e
 cargo build --release
 
 while read d; do
-  if [ -d "$d" ]; then
+  if [ -d "$d" ] && [ -f "$d/a.xml" ]; then
     td=`mktemp -d`
     GOBLINT=yes ./scripts/test_dir.sh $d $td
     rm -rf $td
